Added additional dependencies. Reconfigure server apart from index.js so that testing and local server can run.
