+++ package.json
@@ -8,6 +8,7 @@
   "license": "MIT",
   "dependencies": {
     "cors": "^2.8.5",
+    "dotenv": "^7.0.0",
     "express": "^4.16.4",
     "knex": "^0.16.3",
     "node": "^11.12.0",
